+++ classes/Transport/Smtp2goSmtpTransport.php
@@ -8,7 +8,7 @@ use Symfony\Component\Mailer\Transport\Smtp\EsmtpTransport;
 
 class Smtp2goSmtpTransport extends EsmtpTransport
 {
-    public function __construct(string $username, string $password, int $port = 2525, bool $tls = false, EventDispatcherInterface $dispatcher = null, LoggerInterface $logger = null)
+    public function __construct(string $username, string $password, int $port = 2525, bool $tls = false, ?EventDispatcherInterface $dispatcher = null, ?LoggerInterface $logger = null)
     {
         parent::__construct('mail.smtp2go.com', $port, $tls, $dispatcher, $logger);